requirements.txt
Flask==3.1.0
flask-caching==2.3.0
flask-limiter==3.8.0
flask-htmx==0.3.2
yfpy==13.0.0
redis==5.2.0
gunicorn==23.0.0
python-dotenv==1.0.0
pydantic==2.10.0
